refactor: Comments 훅 분리
