+++ src/features/comment/hooks/index.ts
@@ -4,3 +4,4 @@ export * from "./useUpdateCommentMutation"
 export * from "./useDeleteCommentMutation"
 export * from "./useUpdateLikeMutation"
 export * from "./useCommentsQuery"
+export * from "./useComments"
